Attempt to create and load a task
A task is currently being attempted to be stored in a SINGLE FILE. The
idea: the file will be loaded and the object loaded (a savepackage) will
be taken out and appended to with the new method setTaskModels().
Currently, there is no error saving although there is a null pointer
exception when the getTaskModels() method is ran when trying to display
the tasks. There are no errors in loading the file although nothing may
be loaded or saved.