Added Notification and fixed (temporarily) view tasks
Notifications are now added and working. To test them, click on the
create task button on the homepage which will display a notification
that when clicked, will take you to the viewTasks page. View tasks has
been temporarily fixed to simply display a blank screen but will be
fixed to say a more meaningful message like "no tasks" later.